+++ src/main/java/com/team/independence/asset/mapper/AssetAccountMapper.java
@@ -1,8 +1,8 @@
 package com.team.independence.asset.mapper;
 
-import com.team.independence.asset.domain.AssetAccount;
-import com.team.independence.asset.dto.AssetAccountDetailItem;
-import com.team.independence.asset.dto.AssetAccountQueryItem;
+import com.team.independence.asset.domain.account.AssetAccount;
+import com.team.independence.asset.dto.account.AssetAccountDetailItem;
+import com.team.independence.asset.dto.account.AssetAccountQueryItem;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -16,6 +16,8 @@ public interface AssetAccountMapper {
     List<AssetAccount> findByConnectedInstitutionId(Long connectedInstitutionId);
     List<AssetAccountQueryItem> findWithInstitutionByMemberId(Long memberId);
     List<AssetAccountDetailItem> findAllDetailsByMemberId(Long memberId);
+    List<AssetAccountDetailItem> findBankAccountDetailsByMemberId(Long memberId);
+    List<AssetAccountDetailItem> findStockAccountDetailsByMemberId(Long memberId);
     Long sumCurrentValueByMemberId(@Param("memberId") Long memberId);
 
     /** 회원이 연동한 자산 계좌 중 주어진 asset_category에 속하는 계좌들의 현재가치 합계(원). 없으면 0. */
