+++ src/main/java/com/team/independence/asset/mapper/AssetSummaryMapper.java
@@ -1,6 +1,6 @@
 package com.team.independence.asset.mapper;
 
-import com.team.independence.asset.domain.AssetSummary;
+import com.team.independence.asset.domain.summary.AssetSummary;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
