+++ src/main/resources/mybatis/mapper/asset/ConnectedAccountMapper.xml
@@ -4,7 +4,7 @@
 
 <mapper namespace="com.team.independence.asset.mapper.ConnectedAccountMapper">
 
-    <resultMap id="connectedAccountMap" type="com.team.independence.asset.domain.ConnectedAccount">
+    <resultMap id="connectedAccountMap" type="com.team.independence.asset.domain.codef.ConnectedAccount">
         <id     property="id"              column="id"/>
         <result property="memberId"        column="member_id"/>
         <result property="connectedId"     column="connected_id"/>
@@ -27,13 +27,13 @@
         SELECT member_id FROM connected_account WHERE connected_status = 'ACTIVE'
     </select>
 
-    <insert id="insert" parameterType="com.team.independence.asset.domain.ConnectedAccount"
+    <insert id="insert" parameterType="com.team.independence.asset.domain.codef.ConnectedAccount"
             useGeneratedKeys="true" keyProperty="id">
         INSERT INTO connected_account (member_id, connected_id, birth_date, connected_status)
         VALUES (#{memberId}, #{connectedId}, #{birthDate}, 'ACTIVE')
     </insert>
 
-    <update id="updateSyncStatus" parameterType="com.team.independence.asset.domain.ConnectedAccount">
+    <update id="updateSyncStatus" parameterType="com.team.independence.asset.domain.codef.ConnectedAccount">
         UPDATE connected_account
            SET sync_status    = #{syncStatus},
                last_synced_at = #{lastSyncedAt}
